TestB/requirements.txt
streamlit
chromadb
langchain
langchain-openai
langchain-groq
langchain-core
langgraph
pydantic
aiohttp
numpy
faiss-cpu
sentence-transformers
openai
python-dotenv